Fixed displayTraderItems to actually print out the items.
Created new method displayTraderItemsTwo, which is the exact same as displayTraderItems, but does not print out the following line:
"Choose the item you want to remove by typing in its respective number:"

NOTE: This "displayTraderItemsTwo" method should be removed before final submission! This is just a hotfix so I can work on TraderSystem, and we should figure out a way to combine these methods sooner rather than later.